requirements.txt
asgiref==3.5.0
black==22.1.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==2.2.3
Django==4.0.2
django-debug-toolbar==3.2.4
django-filter==21.1
djangorestframework==3.13.1
drf-yasg==1.20.0
idna==3.3
importlib-metadata==4.11.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
Markdown==3.3.6
MarkupSafe==2.1.0
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.1
psycopg2==2.9.3
pyparsing==3.0.7
python-decouple==3.6
pytz==2021.3
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.1.1
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.8
zipp==3.7.0
